% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census.R
\name{stl_get_census}
\alias{stl_get_census}
\title{Get Census Variables}
\usage{
stl_get_census(region, level, topic, table, variable, year, product,
  geometry = FALSE)
}
\arguments{
\item{region}{One of \code{"city"}, \code{"city/county"}, \code{"core"}, \code{"ew gateway"},
\code{"metro east"}, \code{"metro west"}, or \code{"full metro"}}

\item{level}{One of...}

\item{topic}{For all users who wish to make common data requests.}

\item{table}{For advanced users who wish to specify a full Decennial Census or American
Comunity Survey table.}

\item{variable}{For advanced users who wish to specify a specific Decennial Census or American
Comunity Survey variable.}

\item{year}{Vintage for data. For 1990, 2000, and 2010, it is possible to obtain data from
the Decennial Census. For 2010-2017, it is possible to obtain data from the American
Community Survey.}

\item{product}{One of \code{"census"} for Decennial Census, \code{"acs1"} for the 1-year
American Community Survey (ACS) estimates, \code{"acs3"} for the 3-year ACS estimates
(only available 2010-2013), or \code{"acs5"} for the 5-year ACS estimates.}

\item{geometry}{If \code{TRUE}, will return spatial data as an \code{sf} object; defaults
to \code{FALSE}.}
}
\description{
Pull data from the U.S. Census Bureau's API via the \code{tidycensus} package.
}
