% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometry.R
\name{stl_get_geo}
\alias{stl_get_geo}
\title{Get Census Spatial Data}
\usage{
stl_get_geo(region, level, year = 2017, crs = 26915, ...)
}
\arguments{
\item{region}{One of \code{"city"}, \code{"city/county"}, \code{"core"}, \code{"ew gateway"},
\code{"metro east"}, \code{"metro west"}, or \code{"full metro"}}

\item{level}{One of...}

\item{year}{Vintage for data; defaults to 2017.}

\item{crs}{EPSG code to transform data to; defaults to UTM 15N.}

\item{...}{Advanced options to pass on to the \code{tigris} package.}
}
\description{
Pull TIGER/Line data from the U.S. Census Bureau's API via the
    \code{tidycensus} package. These are commonly known as "shapefiles"
    in GIS contexts.
}
