% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_access.R
\name{stl_install_keys}
\alias{stl_install_keys}
\title{Install API Keys for Census and BLS APIs}
\usage{
stl_install_keys(census, bls, overwrite = FALSE)
}
\arguments{
\item{census}{Your Census Bureau API key, can be obtained from \href{Census Bureau API}{} If you already have a \code{tidycensus} key installed, you do not need this}

\item{bls}{Bureau of Labor Statistics API key, can be obtained from \href{Bureau of Labor Statistics}{} If you already have a \code{blscrapeR} key installed you do not need this}

\item{overwrite}{Logical, if TRUE will overwrite the installation in the .Renviron of the keys specified}
}
\description{
Install API Keys for Census and BLS APIs
}
